Game files
This game is really fun to play, it requires four people to play. Makes your free time worth it. To use, Clone this git and unzip the data.zip which has the assets and keep the unzipped folder in the same folder as of Chauka_Bara.py. Have lots of FUN..... : )